Refaktoringas: sumažinti komponentų dydį, sumažinti props "pertekliaus" kiekį per RouteProvider
RouteDetails.tsx buvo per didelis (332 eil.) ir NavigationBanner rodė du
vizualiai atskirus elementus viename komponente. Pertvarkyta struktūra:

- RouteProvider dabar registruoja papildomus context'us (useRouteResult,
  useNavigationProgress), analogiškai kaip MapProvider turi useMapConfig/
  useMapSelection/useMapTransform. Ten pat perkelta useRouting/
  useRouteProgress iškvietimai bei off-route/arrival logika, kurios anksčiau
  buvo RouteManager.
- RouteManager sutrumpėjo iki paprasto 3 komponentų surinkimo be jokių props
  (68 → 14 eilučių).
- RouteLayer/RouteDetails nebegauna routeResult/progress per props - skaito
  tiesiai iš RouteProvider context'ų.
- NavigationBanner (109 eil., 2 vizualiai atskiri elementai) padalintas į
  NavigationInstructionBanner (viršutinė juosta) ir NavigationProgressBar
  (apatinė juosta).
- Navigacijos mobili UI (kompaktiška juosta + pilnas sąrašas) iškelta iš
  RouteDetails į naują MobileNavigationView komponentą, kurį RouteManager
  atvaizduoja kaip atskirą brolinį elementą.
- Bendra "aktyvios instrukcijos" skaičiavimo logika (steps/activeInstruction/
  liveDistanceToNext/remainingTime) iškelta į naują hook'ą
  useActiveInstruction, naudojamą ir RouteDetails, ir MobileNavigationView.
- Žemėlapio fokusavimo logika (flyTo į instrukciją/tašką) iškelta į
  useRouteMapFocus, pašalinant dubliuotą kodą tarp RouteDetails ir
  MobileNavigationView.

RouteDetails.tsx: 332 → 178 eilučių. Jokių funkcinių pakeitimų - patikrinta
Playwright regresija (progresas, pauzė/tęsimas, atvykimas, GPS tikslumas,
nuklydimo perskaičiavimas, mobili/desktop navigacija).
